requirements.txt
pytest>=7.0
lxml>=4.9
python-dateutil>=2.8
PyYAML>=6.0
tabulate>=0.9
flake8>=6.0
